inline untyped_plutus fns
